Add plot module
